Cleanup of git repo and addition of gitignore
