+++ .gitignore
@@ -1,0 +1,4 @@
+#gitignore
+.o
+.vscode
+.DS_Store
