--- hw1/memory.txt
@@ -1,256 +1,0 @@
-7F
-24
-7A
-41
-5E
-7F
-2B
-47
-33
-32
-53
-7F
-78
-67
-2E
-7F
-68
-7F
-7F
-60
-7F
-7F
-77
-7F
-7F
-6D
-55
-7F
-7F
-7F
-60
-7F
-7F
-7F
-7F
-4F
-12
-2E
-18
-76
-02
-7F
-7F
-00
-08
-35
-74
-20
-02
-7F
-5A
-7F
-6F
-6F
-0D
-0D
-17
-7F
-3D
-7F
-7F
-7F
-7D
-38
-3A
-7F
-7F
-59
-76
-7F
-7F
-29
-7F
-7F
-6E
-7F
-64
-2E
-7F
-7F
-54
-7F
-7F
-7F
-20
-3B
-06
-7F
-1C
-68
-7F
-7F
-5E
-35
-70
-7F
-20
-78
-7F
-0B
-7F
-7F
-48
-22
-7F
-7F
-7F
-7F
-31
-7F
-58
-6B
-28
-7F
-7F
-7F
-7F
-11
-7F
-7F
-7F
-7F
-7F
-28
-75
-7F
-7F
-7F
-7F
-7F
-7F
-7F
-09
-7F
-68
-09
-7F
-27
-25
-7F
-41
-53
-66
-68
-62
-7F
-2B
-30
-18
-52
-7F
-3C
-7F
-7F
-7F
-7F
-4B
-4F
-7F
-53
-7F
-7F
-7F
-43
-7F
-3C
-74
-62
-7F
-7F
-7F
-7F
-3C
-7F
-0D
-7F
-7F
-7F
-7F
-7F
-68
-00
-7F
-35
-38
-53
-18
-7F
-7F
-7F
-56
-7F
-7F
-7F
-7F
-5B
-7F
-58
-7F
-7F
-7F
-20
-7F
-06
-6A
-7F
-7F
-5E
-7F
-7F
-3E
-21
-39
-59
-49
-7F
-0D
-7F
-2A
-7F
-2B
-7F
-7F
-5B
-0C
-36
-65
-55
-3B
-7F
-7F
-7F
-70
-7F
-01
-7F
-7F
-3B
-7F
-7F
-7F
-7F
-7F
-54
-39
-50
-7F
-7F
-4A
-67
-7F
-7F
-7F
-34
-7F
-7F
