+++ memory.h
@@ -2,7 +2,7 @@
 // CSCI 340 - Operating Systems
 // Fall 2018
 // memory.h header file
-// 
+//
 // Homework 1
 //
 // -----------------------------------
@@ -18,7 +18,7 @@
 // DEBUG FLAGS
 #define MEM_DEBUG 0
 
-// Simulator Parameters 
+// Simulator Parameters
 #define NUM_BLOCK_OFFSET_BITS 2
 #define MAX_SIZE 1000
 
@@ -29,7 +29,9 @@
 // --------------------------------
 unsigned int* phy_memory;
 unsigned int* block_location;
+unsigned int* addresses;
 unsigned int addr_bits;
+unsigned int num_addresses;
 
 
 // --------------------------------
@@ -40,9 +42,9 @@ unsigned int addr_bits;
 // Function prototype that creates the
 // phy_memory array by reading the bytes
 // defined in the memory.txt file.
-// 
 //
-// Arguments:	character array (i.e. string) 
+//
+// Arguments:	character array (i.e. string)
 //
 // Return:     	Number of addressable bits (no error)
 //		READ_ERROR (error)
@@ -53,9 +55,9 @@ unsigned int readMemoryFile( char* file_name );
 
 // ------------------------------------
 // Function prototype that computes the
-// number of blocks given the number of 
+// number of blocks given the number of
 // addressable bits.
-// 
+//
 //
 // Arguments:	number of addressable bits
 //		number of block offset bits
@@ -67,10 +69,10 @@ unsigned int readMemoryFile( char* file_name );
 int numberOfBlocks( unsigned int addr_bits, unsigned int num_block_offset_bits );
 
 // ------------------------------------
-// Function prototype that identifies the 
-// starting physcial memory address location 
+// Function prototype that identifies the
+// starting physcial memory address location
 // for each block.
-// 
+//
 //
 // Arguments:	number of blocks
 //		number of block offset bits
