--- ubuntu-xenial-16.04-cloudimg-console.log
@@ -1,43 +1,0 @@
-[ 5291.032741] invalid opcode: 0000 [#1] SMP 
-[ 5291.034930] Modules linked in: vboxsf isofs crct10dif_pclmul crc32_pclmul ghash_clmulni_intel input_leds serio_raw vboxguest video ib_iser rdma_cm iw_cm ib_cm ib_sa ib_mad ib_core ib_addr iscsi_tcp libiscsi_tcp libiscsi scsi_transport_iscsi autofs4 btrfs raid10 raid456 async_raid6_recov async_memcpy async_pq async_xor async_tx xor raid6_pq libcrc32c raid1 raid0 multipath linear aesni_intel aes_x86_64 glue_helper lrw gf128mul ablk_helper cryptd mptspi scsi_transport_spi mptscsih psmouse mptbase e1000
-[ 5291.167604] CPU: 1 PID: 0 Comm: swapper/1 Not tainted 4.4.0-134-generic #160-Ubuntu
-[ 5291.175063] Hardware name: innotek GmbH VirtualBox/VirtualBox, BIOS VirtualBox 12/01/2006
-[ 5291.184300] task: ffff88003d858e00 ti: ffff88003d864000 task.ti: ffff88003d864000
-[ 5291.192161] RIP: 0010:[<ffffffff8153573b>]  [<ffffffff8153573b>] add_interrupt_randomness+0x14b/0x1e0
-[ 5291.210261] RSP: 0018:ffff88003fd03e90  EFLAGS: 00010082
-[ 5291.220967] RAX: 0000000000000000 RBX: ffff88003fd14ee0 RCX: 000000000000000a
-[ 5291.280925] RDX: ffffffff822008c8 RSI: ffff88003fd14ef0 RDI: ffffffff81eca1e0
-[ 5291.302305] RBP: ffff88003fd03ec0 R08: 000000000000004a R09: 000000000000007f
-[ 5291.304582] R10: 00000000d69650c3 R11: 0000000000000068 R12: 00000001001307f3
-[ 5291.313342] R13: ffffffff81eca1e0 R14: ffffffff81eca228 R15: 0000000000000000
-[ 5291.372617] FS:  0000000000000000(0000) GS:ffff88003fd00000(0000) knlGS:0000000000000000
-[ 5291.386868] CS:  0010 DS: 0000 ES: 0000 CR0: 0000000080050033
-[ 5291.393233] CR2: 00007ffbdbc64000 CR3: 000000003ac44000 CR4: 0000000000060670
-[ 5291.401467] Stack:
-[ 5291.404149]  ffff88003fd03ea8 d3a46e27b3014d4e 0000000000000000 0000000000000013
-[ 5291.412231]  0000000000000001 0000000000000001 ffff88003fd03f10 ffffffff810df9ed
-[ 5291.421683]  ffff880036fef200 ffffffff81f3dd40 0000008000000000 ffff880036fef200
-[ 5291.430967] Call Trace:
-[ 5291.433866]  <IRQ> [ 5291.434951] e1000: enp0s3 NIC Link is Down
-
-[ 5291.440791]  [<ffffffff810df9ed>] handle_irq_event_percpu+0x15d/0x1e0
-[ 5291.449040]  [<ffffffff810dfaae>] handle_irq_event+0x3e/0x60
-[ 5291.456029]  [<ffffffff810e2e9c>] handle_fasteoi_irq+0x9c/0x160
-[ 5291.458713]  [<ffffffff81031143>] handle_irq+0x23/0x30
-[ 5291.460801]  [<ffffffff818587e3>] do_IRQ+0x53/0xe0
-[ 5291.462893]  [<ffffffff81855ebf>] common_interrupt+0xbf/0xbf
-[ 5291.465181]  <EOI> 
-[ 5291.472405]  [<ffffffff81038f80>] ? idle_notifier_unregister+0x20/0x20
-[ 5291.482157]  [<ffffffff810667c6>] ? native_safe_halt+0x6/0x10
-[ 5291.483936]  [<ffffffff81038f9e>] default_idle+0x1e/0xe0
-[ 5291.485971]  [<ffffffff81039b65>] arch_cpu_idle+0x15/0x20
-[ 5291.487737]  [<ffffffff810c8bba>] default_idle_call+0x2a/0x40
-[ 5291.490062]  [<ffffffff810c8f27>] cpu_startup_entry+0x2f7/0x360
-[ 5291.503528]  [<ffffffff81052e07>] start_secondary+0x177/0x1b0
-[ 5291.505624] Code: a0 ec 81 4c 0f 45 e8 4d 8d 75 48 4c 89 f7 e8 0d f7 31 00 85 c0 74 b7 4c 89 63 10 ba 10 00 00 00 48 89 de 4c 89 ef e8 a5 f0 ff ff <48> 0f c7 f8 0f 92 c2 84 d2 48 89 45 d0 be 01 00 00 00 74 16 48 
-[ 5291.549141] RIP  [<ffffffff8153573b>] add_interrupt_randomness+0x14b/0x1e0
-[ 5291.561187]  RSP <ffff88003fd03e90>
-[ 5291.571068] ---[ end trace ca2079f2c09cd897 ]---
-[ 5291.580745] Kernel panic - not syncing: Fatal exception in interrupt
-[ 5291.585081] Kernel Offset: disabled
-[ 5291.589655] ---[ end Kernel panic - not syncing: Fatal exception in interrupt
